+++ requests_retry_session/requests_retry_session.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2022-2025 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2022-2026 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -21,7 +21,6 @@
 # ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
 #
-
 """
 Return a requests session with retries, timeouts, and logging.
 
@@ -34,15 +33,18 @@ Created on Nov 2, 2020
 @author: jsl
 """
 
-# Because we want to support Python 3.6 and 3.9, use old-style type hint syntax
+import sys
 from typing import Optional, Tuple
-from typing_extensions import TypedDict, Unpack
+
+if sys.version_info >= (3, 9):
+    from typing import TypedDict
+else:
+    from typing_extensions import TypedDict
 
 import requests
 
 from .timeout_http_adapter import TimeoutHTTPAdapter
 from .retry_with_logs import RetryWithLogs
-
 
 DEFAULT_BACKOFF_FACTOR = 0.5
 DEFAULT_CONNECT_TIMEOUT = 3
@@ -51,7 +53,12 @@ DEFAULT_READ_TIMEOUT = 10
 DEFAULT_RETRIES = 10
 DEFAULT_STATUS_FORCELIST = (500, 502, 503, 504)
 
+
 class RequestsRetryAdapterArgs(TypedDict, total=False):
+    """
+    Used to represent the parameters to the requests_retry_adapter function.
+    Helpful with type hinting.
+    """
     retries: int
     backoff_factor: float
     status_forcelist: Tuple[int, ...]
@@ -59,9 +66,15 @@ class RequestsRetryAdapterArgs(TypedDict, total=False):
     read_timeout: int
 
 
-def requests_session(adapter: requests.adapters.HTTPAdapter,
-                     session: Optional[requests.Session] = None,
-                     protocol: str = DEFAULT_PROTOCOL) -> requests.Session:
+def requests_session(adapter,
+                     session = None,
+                     protocol = DEFAULT_PROTOCOL):
+    """
+    adapter: requests.adapters.HTTPAdapter,
+    session: Optional[requests.Session]
+    protocol: str
+    -> requests.Session:
+    """
     session = session or requests.Session()
     # Must mount to http://
     # Mounting to only http will not work!
@@ -69,11 +82,20 @@ def requests_session(adapter: requests.adapters.HTTPAdapter,
     return session
 
 
-def requests_retry_adapter(retries: int = DEFAULT_RETRIES,
-                           backoff_factor: float = DEFAULT_BACKOFF_FACTOR,
-                           status_forcelist: Tuple[int, ...] = DEFAULT_STATUS_FORCELIST,
-                           connect_timeout: int = DEFAULT_CONNECT_TIMEOUT,
-                           read_timeout: int = DEFAULT_READ_TIMEOUT) -> TimeoutHTTPAdapter:
+def requests_retry_adapter(
+        retries = DEFAULT_RETRIES,
+        backoff_factor = DEFAULT_BACKOFF_FACTOR,
+        status_forcelist = DEFAULT_STATUS_FORCELIST,
+        connect_timeout = DEFAULT_CONNECT_TIMEOUT,
+        read_timeout = DEFAULT_READ_TIMEOUT):
+    """
+    retries: int
+    backoff_factor: float
+    status_forcelist: Tuple[int, ...]
+    connect_timeout: int
+    read_timeout: int
+    -> .timeout_http_adapter.TimeoutHTTPAdapter:
+    """
     retry = RetryWithLogs(
         total=retries,
         read=retries,
@@ -81,11 +103,22 @@ def requests_retry_adapter(retries: int = DEFAULT_RETRIES,
         backoff_factor=backoff_factor,
         status_forcelist=status_forcelist,
     )
-    return TimeoutHTTPAdapter(max_retries=retry, timeout=(connect_timeout, read_timeout))
+    return TimeoutHTTPAdapter(max_retries=retry,
+                              timeout=(connect_timeout, read_timeout))
 
 
-def requests_retry_session(session: Optional[requests.Session] = None,
-                           protocol: str = DEFAULT_PROTOCOL,
-                           **adapter_kwargs: Unpack[RequestsRetryAdapterArgs]) -> requests.Session:
+def requests_retry_session(
+        session = None,
+        protocol = DEFAULT_PROTOCOL,
+        **adapter_kwargs
+):
+    """
+    session: Optional[requests.Session]
+    protocol: str
+    **adapter_kwargs: Unpack[RequestsRetryAdapterArgs]
+    -> requests.Session
+    """
     adapter = requests_retry_adapter(**adapter_kwargs)
-    return requests_session(adapter=adapter, session=session, protocol=protocol)
+    return requests_session(adapter=adapter,
+                            session=session,
+                            protocol=protocol)