+++ requests_retry_session/retry_session_manager.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2024-2025 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2024-2026 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -23,14 +23,10 @@
 #
 
 from contextlib import closing, contextmanager, AbstractContextManager
-from typing import Iterator, Optional, TYPE_CHECKING
-from typing_extensions import Self, Unpack
-
 import requests
 
 from .requests_retry_session import requests_retry_adapter, requests_session, \
-                                    RequestsRetryAdapterArgs, DEFAULT_PROTOCOL
-
+                                    DEFAULT_PROTOCOL
 
 class RetrySessionManager(AbstractContextManager):
     """
@@ -38,42 +34,72 @@ class RetrySessionManager(AbstractContextManager):
     retry session only when needed, and to clean it up in their __exit__ function.
     This class is not thread safe.
     """
-    def __init__(self, protocol: Optional[str]=None,
-                 **adapter_kwargs: Unpack[RequestsRetryAdapterArgs]):
+
+    def __init__(self,
+                 protocol = None,
+                 **adapter_kwargs):
+        """
+        protocol: Optional[str]
+        **adapter_kwargs: Unpack[.requests_retry_session.RequestsRetryAdapterArgs]
+        -> None
+        """
+        # self._requests_adapter: Optional[.timeout_http_adapter.TimeoutHTTPAdapter]
         self._requests_adapter = None
+        # self._requests_session: Optional[requests.Session]
         self._requests_session = None
+        # self._requests_protocol: str
         self._requests_protocol = protocol if protocol is not None else DEFAULT_PROTOCOL
+        # self._requests_retry_adapter_kwargs: .requests_retry_session.RequestsRetryAdapterArgs
         self._requests_retry_adapter_kwargs = adapter_kwargs
 
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
+    def __exit__(  # pylint: disable=useless-return
+            self, exc_type,
+            exc_val,
+            exc_tb):
+        """
+        exc_type: Optional[Type[BaseException]],
+        exc_val: Optional[BaseException],
+        exc_tb: Optional[TracebackType]) -> Optional[bool]:
+        -> Optional[bool]
+        """
         if self._requests_session is not None:
             self._requests_session.close()
-            self._requests_adapter.close()
             self._requests_session = None
+        if self._requests_adapter is not None:
+            self._requests_adapter.close()
             self._requests_adapter = None
-
-    # The following is needed to work around https://github.com/python/typing/issues/1992
-    if TYPE_CHECKING:
-        def __enter__(self) -> Self:
-            return self
+        # The following return statement is not needed, but it makes mypy sad without it
+        return None
 
     @property
-    def requests_session(self) -> requests.Session:
+    def requests_session(self):
+        """
+        Returns the requests retry session, after initializing it if needed
+        -> requests.Session
+        """
         if self._requests_session is None:
-            self._requests_adapter = requests_retry_adapter(**self._requests_retry_adapter_kwargs)
-            self._requests_session = requests_session(adapter=self._requests_adapter,
-                                                      protocol=self._requests_protocol)
+            self._requests_adapter = requests_retry_adapter(
+                **self._requests_retry_adapter_kwargs)
+            self._requests_session = requests_session(
+                adapter=self._requests_adapter,
+                protocol=self._requests_protocol)
         return self._requests_session
 
 
 @contextmanager
-def retry_session_manager(protocol: Optional[str]=None,
-                          **adapter_kwargs: Unpack[RequestsRetryAdapterArgs]) -> Iterator[None]:
+def retry_session_manager(
+    protocol = None,
+    **adapter_kwargs
+):
     """
     Provides a context manager that will clean up both the session and the adapter on exit
+
+    protocol: Optional[str]
+    **adapter_kwargs: Unpack[.requests_retry_session.RequestsRetryAdapterArgs]
+    -> Iterator[requests.Session]
     """
     requests_protocol = protocol if protocol is not None else DEFAULT_PROTOCOL
     with closing(requests_retry_adapter(**adapter_kwargs)) as adapter:
-        with requests_session(adapter=adapter, protocol=requests_protocol) as session:
+        with requests_session(adapter=adapter,
+                              protocol=requests_protocol) as session:
             yield session