+++ requests_retry_session/retry_session_manager.py
@@ -39,6 +39,8 @@ class RetrySessionManager(AbstractContextManager):
                  protocol = None,
                  **adapter_kwargs):
         """
+        If specified, protocol should omit the trailing "://" because it will be automatically appended later
+
         protocol: Optional[str]
         **adapter_kwargs: Unpack[.requests_retry_session.RequestsRetryAdapterArgs]
         -> None
@@ -93,6 +95,8 @@ def retry_session_manager(
 ):
     """
     Provides a context manager that will clean up both the session and the adapter on exit
+
+    If specified, protocol should omit the trailing "://" because it will be automatically appended later
 
     protocol: Optional[str]
     **adapter_kwargs: Unpack[.requests_retry_session.RequestsRetryAdapterArgs]