+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added `mypy` and `pylint` at build time
+
 ## [0.5.5] - 2026-04-14
 
 ### Changed