+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Backported improvements from 1.0 branch, to keep v0.5 and v1.0 as consistent as possible,
   except for the type annotations
+- Changed type hint for timeout arguments from `int` to `float` (`float` values
+  already work fine)
 
 ### Removed
 - Removed `py.typed` from package; accurate type annotations will only be for versions of this